pyproject.toml
[project]
name = "fastapi_backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=13.3.0",
    "aiobotocore==2.16.0",
    "aiohappyeyeballs==2.4.4",
    "aiohttp==3.11.11",
    "aioitertools==0.12.0",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.7.0",
    "attrs==24.3.0",
    "boto3==1.35.78",
    "botocore==1.35.78",
    "certifi==2024.12.14",
    "click==8.1.7",
    "dnspython==2.7.0",
    "duckdb==1.1.3",
    "email-validator==2.2.0",
    "fastapi==0.115.6",
    "fastapi-cli==0.0.6",
    "frozenlist==1.5.0",
    "gevent==24.11.1",
    "gotrue>=2.12.3",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.4",
    "jmespath==1.0.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "numpy==2.2.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "propcache==0.2.1",
    "pydantic==2.10.3",
    "pydantic-core==2.27.1",
    "pygments==2.18.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.19",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "rich==13.9.4",
    "rich-toolkit==0.12.0",
    "s3transfer==0.10.4",
    "setuptools==75.6.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "structlog==24.4.0",
    "supabase>=2.13.0",
    "typer==0.15.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "urllib3==2.2.3",
    "uvicorn==0.32.1",
    "uvloop==0.21.0",
    "watchfiles==1.0.3",
    "websockets==14.1",
    "wrapt==1.17.0",
    "yarl==1.18.3",
    "zope-event==5.0",
    "zope-interface==7.2",
]
